% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{funr-package}
\alias{funr}
\alias{funr-package}
\title{funr: Functional Programming with Parallelism in R}
\description{
This package provides a consistent and intuitive API for applying functions
over lists, vectors, and data frames, with built-in support for parallelism
and progress bars.
}
\author{
\strong{Maintainer}: Imad EL BADISY \email{elbadisyimad@gmail.com}

}
\keyword{internal}
