% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{functionals-package}
\alias{functionals}
\alias{functionals-package}
\title{functionals: Functional programming with parallelism and progress track in R}
\description{
This package provides a consistent and intuitive API for applying functions
over lists, vectors, and data frames, with built-in support for parallelism
and progress bars.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://CRAN.R-project.org/package=functionals}
  \item Report bugs at \url{https://github.com/ielbadisy/functionals/issues}
}

}
\author{
\strong{Maintainer}: Imad El Badisy \email{elbadisyimad@gmail.com}

}
\keyword{internal}
